Importer: Dispatchify `mapAuthors()`
Part of #5302

The importer is being reduxified gradually. The current action creators
all use the `Dispatcher` from Facebook in a tightly integrated way. The
goal of this (and future related) PRs is to try and rid these action
creators from the `Dispatcher` so that they can be ready for Redux
without making a massive changing PR.

The structural change is meant to bridge into using `connect()` with the
store dispatch. Currently it looks odd to pull in the `Dispatcher`
inside a React component, but my hope is that this will foster the
gradual change and will only exist for a short while.

There are no functional or stylistic changes in this patch.